% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{MoonShineR-package}
\alias{MoonShineR}
\alias{MoonShineR-package}
\title{MoonShineR Package}
\description{
Predict and plot natural light intensity from moonlight, sunlight, and twilight.
}
\author{
\strong{Maintainer}: Lok Poon \email{poonchiulok@gmail.com} (\href{https://orcid.org/0000-0003-3547-1614}{ORCID})

}
\keyword{internal}
