% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{MoonShineR}
\alias{MoonShineR-package}
\alias{MoonShineR}
\title{MoonShineR Package}
\description{
Description of what the package does.
}
\author{
\strong{Maintainer}: Lok Poon \email{poonchiulok@gmail.com} (\href{https://orcid.org/0000-0003-3547-1614}{ORCID})

}
