% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_lux.R
\name{predict_lux}
\alias{predict_lux}
\title{A function to predict ground illuminance}
\usage{
predict_lux(
  latitude = NULL,
  longitude = NULL,
  site_elev = 0,
  time_zone = NULL,
  date_start = NULL,
  time_start = "00:00:00",
  duration_day = NULL,
  time_interval_minutes = 5,
  darksky_value = 8e-04,
  illuminance_type_plot = "moon_final_lux_nighttime",
  output_directory = NULL,
  save_table = FALSE,
  save_plot = FALSE,
  plot_width = 11,
  plot_height = 8.5,
  plot_y_max = "AUTO",
  plot_dayttime_gray_mask = TRUE
)
}
\arguments{
\item{latitude}{\code{numeric}. Latitude in decimal degrees (e.g., -4.21528).}

\item{longitude}{\code{numeric}. Longitude in decimal degrees (e.g., -69.94056).}

\item{site_elev}{\code{numeric}. Site elevation in meters (e.g., 0 = sea level). Default = 0. Elevation correction only applies to moonlight but not sunlight and twilight. Site elevation of a coordinate location can be obtained from \url{https://www.dcode.fr/earth-elevation}.}

\item{time_zone}{\code{character}. Time zone for the location set (e.g., “EST”). Remember to change time_zone to corrospond it to the location set. For a list of time zone names, enter OlsonNames(tzdir = NULL) in R console. Use a time zone without DST (e.g., use "EST" instead of "America/New_York").}

\item{date_start}{\code{character}. Starting date of the simulation ("YYYY-MM-DD").}

\item{time_start}{\code{character}. Starting time of the simulation ("hh:mm:ss"). Default = "00:00:00".}

\item{duration_day}{\code{numeric}. Duration of the simulation in days.}

\item{time_interval_minutes}{\code{numeric}.The temporal resolution of the simulation in minutes. E.g., 5 = calculates the illuminance every 5 minutes. Using small time interval requires longer computation time.}

\item{darksky_value}{\code{numeric}. A baseline illumination added to the model to represent other constant nocturnal light sources (e.g., starlight and airglow). Default = 0.0008. Change it to zero if a completely dark sky is preferred.}

\item{illuminance_type_plot}{\code{character}. Choose one type of illuminance to plot. See options in the section below. Default = "moon_final_lux_nighttime".}

\item{output_directory}{\code{character}. Directory to save the output table (.csv) and plot (.pdf). NULL if saving is turned OFF (i.e., save_table = FALSE and save_plot = FALSE).}

\item{save_table}{\code{logical}. TRUE to save output .csv table to the output_directory. FALSE to disable. Default = FALSE.}

\item{save_plot}{\code{logical}. TRUE to save output .pdf plot to the output_directory. FALSE to disable. Default = FALSE.}

\item{plot_width}{\code{numeric}. PDF width in inch. Default = 11.}

\item{plot_height}{\code{numeric}. PDF height in inch. Default = 8.5.}

\item{plot_y_max}{\code{character} "AUTO", or \code{numeric}. Let the plot y-axis scale automatically or manually set a y-axis upper limit. Default = "AUTO".}

\item{plot_dayttime_gray_mask}{'logical`. TRUE to mask daytime plot line in gray. FALSE to disable (plot line always black).}
}
\description{
\itemize{
\item This function predicts moonlight, sunlight, and/or twilight ground illumination in lux for any defined geographical location and time period. It creates a data.frame output and automatically plot to the console. Automatic saving of the table (.csv) and plot (.pdf) is optional.
\item To learn more about MoonShine, see instruction manual: \url{https://lokpoon.github.io/moonshine_manual/overview.html}
}
}
\section{illuminance_type_plot options:}{
Note: these terms corresponding to the columns headers in the output table:
\itemize{
\item \strong{"moon_final_lux"} plots only the illuminance of moonlight (plus the darksky_value) during both day and night.
\item \strong{"moon_final_lux_nighttime"} plots only the illuminance of moonlight at night (no value during daytime, when sun altitude > 0 degrees)
\item \strong{"moonlight_twilight_nighttime"} plots the illuminance of moonlight plus twilight (no value during daytime, when sun altitude > 0 degrees)
\item \strong{"twilight"} plots only the illuminance of twilight (defined as the light when sun altitude < 0 degrees).
\item \strong{"sunlight"} plots only the illuminance of sunlight (defined as the light when sun altitude > 0 degrees).
\item \strong{"sunlight_twilight"} plots only the sum of sunlight and twilight
\item \strong{"total_illuminance_all"} plots all illuminance together, calculated as the sum of moonlight, twilight, and sunlight.
}
}

\examples{
moonlight_output <- predict_lux(latitude = -4.21528, longitude = -69.94056, site_elev = 0, time_zone = "EST", 
                    date_start = "2023-02-27", time_start = "18:00:00", duration_day = 14, time_interval_minutes = 5,
darksky_value = 0.0008, illuminance_type_plot = "moon_final_lux_nighttime",
output_directory = NULL, save_table = FALSE, save_plot = FALSE,
plot_width = 11, plot_height = 8.5, plot_y_max = "AUTO",  plot_dayttime_gray_mask = TRUE)
}
\keyword{moonlight}
