% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_lux.R
\name{predict_lux}
\alias{predict_lux}
\title{Predict moonlight, sunlight, and twilight ground illuminance}
\usage{
predict_lux(
  latitude = NULL,
  longitude = NULL,
  site_elev = 0,
  time_zone = NULL,
  date_start = NULL,
  time_start = "00:00:00",
  duration_day = NULL,
  time_interval_minutes = 5,
  darksky_value = 8e-04,
  output_directory = NULL,
  export_table = FALSE
)
}
\arguments{
\item{latitude}{\code{numeric}. Latitude in decimal degrees (e.g., \code{-4.21528}).}

\item{longitude}{\code{numeric}. Longitude in decimal degrees (e.g., \code{-69.94056}).}

\item{site_elev}{\code{numeric}. Site elevation in meters (e.g., \code{0} is sea level). Default is \code{0}. Elevation correction only applies to moonlight but not sunlight and twilight. Site elevation of a coordinate location can be obtained from \url{https://www.dcode.fr/earth-elevation}.}

\item{time_zone}{\code{character}. Time zone for the location set (e.g., \verb{“EST”}). Remember to change time_zone to correspond it to the location set. For a list of time zone names, enter \code{OlsonNames(tzdir = NULL)} in R console. Use a time zone without DST to avoid confusion (e.g., use \code{"EST"} instead of \code{"America/New_York"}).}

\item{date_start}{\code{character}. Starting date of the simulation (\code{"YYYY-MM-DD"}).}

\item{time_start}{\code{character}. Starting time of the simulation (\code{"hh:mm:ss"}). Default is \code{"00:00:00"}.}

\item{duration_day}{\code{numeric}. Duration of the simulation in days.}

\item{time_interval_minutes}{\code{numeric}.The temporal resolution of the simulation in minutes. E.g., \code{5} calculates the illuminance every 5 minutes. Using small time interval requires longer computation time. Default is \code{5}.}

\item{darksky_value}{\code{numeric}. A baseline illuminance (in lux) added to the model to represent other constant nocturnal light sources (e.g., starlight and airglow). Default is \code{0.0008}. Change it to zero if a completely dark sky is preferred.}

\item{output_directory}{\code{character}. Directory to save the output table (.csv) and plot (.pdf). Ignore output_directory if the export options are turned OFF (i.e., \code{export_table = FALSE}).}

\item{export_table}{\code{logical}. \code{TRUE} to export output .csv table to the output_directory. \code{FALSE} to disable. Default is \code{FALSE}.}
}
\description{
\itemize{
\item predict_lux() predicts moonlight, sunlight, and/or twilight ground illumination (in lux) for any defined geographical location and time period. It creates a data.frame output and automatically plot to the console. Automatic export of the table (.csv) is optional.
\item The presence of lunar eclipse during the simulated period is also reported as a console message. The illuminance reduction during lunar eclipse is not modeled.
\item To create a moonlight or sunlight LED schedule for light re-creation using MoonShineP, user need to download the R script version of the MoonShineR: Moonlight scheduler or MoonShineR: Sunlight/twilight scheduler from GitHub repository: \url{https://github.com/Crampton-Lab/MoonShine}
\item The R script version also have other simulation features, such as simulating the blocking of direct moonlight from horizon obstructions (e.g., tree line or ridgeline), and attenuation by random passing clouds.
\item To learn more about MoonShineR, see instruction manual: \url{https://lokpoon.github.io/moonshine_manual/overview.html}
}
}
\section{Columns found in the output data.frame/.csv table:}{
\itemize{
\item A series of astronomical and illuminance values are reported for every time stamp. In explanation,
\item \strong{datetime} \code{POSIXct}. Datetime in \code{"YYYY-MM-DD hh:mm:ss"}.
\item \strong{phase_angle} \code{numeric}. Phase angle of the moon in degree angle.
\item \strong{fraction} \code{numeric}. Illuminated fraction of the moon.
\item \strong{Z_moon} \code{numeric}. Zenith distance of the moon in degree angle (i.e., the angle of separation from directly overhead). > 90 means the moon is below the horizon.
\item \strong{distance} \code{numeric}. The moon-Earth distance in km.
\item \strong{sun_altitude} \code{numeric}. The sun altitude (relative to the horizon) in degree angle. Negative value means the sun is below the horizon.
\item \strong{"moon_final_lux"} \code{numeric}. Only the illuminance of moonlight (plus the darksky_value) during both day and night.
\item \strong{"moon_final_lux_nighttime"} \code{numeric}. Only the illuminance of moonlight at night (no value during daytime, when \code{sun_altitude} > 0 degrees)
\item \strong{"moonlight_twilight_nighttime"} \code{numeric}. The illuminance of moonlight plus twilight (no value during daytime, when \code{sun_altitude} > 0 degrees)
\item \strong{"twilight"} \code{numeric}. Only the illuminance of twilight (defined as the light when \code{sun_altitude} < 0 degrees).
\item \strong{"sunlight"} \code{numeric}. Only the illuminance of sunlight (defined as the light when \code{sun_altitude} > 0 degrees).
\item \strong{"sunlight_twilight"} \code{numeric}. The sum of sunlight and twilight.
\item \strong{"total_illuminance_all"} \code{numeric}. All illuminance together, calculated as the sum of moonlight, twilight, and sunlight.
\item \strong{"eclipse"} \code{logical}. The presence of a lunar eclipse (all kinds), including penumbral lunar eclipse.
}
}

\section{Attribution}{
\itemize{
\item Astronomical values used in calculating ground illuminance are provided by the R package suncalc (Thieurmel & Elmarhraoui, 2022).
\item Atmospheric pressure at elevation is provided by the R package RPMODEL (Stocker et al., 2020).
}
}

\examples{
# Predict the nighttime moonlight illuminance in Leticia, Colombia,
# for 14 days starting on 2023-02-27 at 6pm.

moonlight_output <- predict_lux(latitude = -4.21528, longitude = -69.94056, site_elev = 0,
                    time_zone = "EST", date_start = "2023-02-27", time_start = "18:00:00",
                    duration_day = 14, time_interval_minutes = 5, darksky_value = 0.0008,
                    output_directory = NULL, export_table = FALSE)
}
\references{
\itemize{
\item Allen, C. W. (1976). Astrophysical quantities. Athelone Press.
\item Austin, R. H., Phillips, B. F., & Webb, D. J. (1976). A method for calculating moonlight illuminance at the earth’s surface. The Journal of Applied Ecology, 13(3), 741.
\item Buratti, B. J., Hillier, J. K., & Wang, M. (1996). The lunar opposition surge: Observations by clementine. Icarus, 124(2), 490–499.
\item Hänel, A., Posch, T., Ribas, S. J., Aubé, M., Duriscoe, D., Jechow, A., Kollath, Z., Lolkema, D. E., Moore, C., Schmidt, N., Spoelstra, H., Wuchterl, G., & Kyba, C. C. M. (2018). Measuring night sky brightness: Methods and challenges. Journal of Quantitative Spectroscopy and Radiative Transfer, 205, 278–290.
\item Krisciunas, K., & Schaefer, B. E. (1991). A model of the brightness of moonlight. Publications of the Astronomical Society of the Pacific, 103, 1033.
\item Laue, E. G. (1970). The measurement of solar spectral irradiance at different terrestrial elevations. Solar Energy, 13(1), 43–57.
\item Schaefer, B. E. (1990). Telescopic limiting magnitudes. Publications of the Astronomical Society of the Pacific, 102, 212.
\item Seidelmann, P. K., United States Naval Observatory, & Great Britain (Eds.). (1992). Explanatory supplement to the Astronomical almanac (Rev. ed.). University Science Books.
\item Stocker, B. D., Wang, H., Smith, N. G., Harrison, S. P., Keenan, T. F., Sandoval, D., Davis, T., & Prentice, I. C. (2020). P-model v1.0: An optimality-based light use efficiency model for simulating ecosystem gross primary production. Geoscientific Model Development, 13(3), 1545–1581.
\item Thieurmel, B., & Elmarhraoui, A. (2022). suncalc: Compute sun position, sunlight phases, moon position and lunar phase. R package version 0.5.1. \url{https://CRAN.R-project.org/package=suncalc}
}
}
\keyword{moonlight}
